requirements.txt
polygon>=1.2.8
massive>=2.0.1
pyyaml>=6.0.2
python-dateutil>=2.9.0.post0
pytz>=2024.1
pandas>=2.2.2
numpy>=2.1.1
pytest>=8.3.3
python-dotenv>=1.0.1
